project structure (including classes and previous modules) was separated by modules; now there are 3 modules in project: all_menu_operators.py, auxilary.py, notes_main.py
