--- menu_operations/auxilary/menu_notes.py
@@ -1,43 +1,0 @@
-
-
-class NotesMenu:
-
-    def show_menu(self):
-        print('1. Распечатать все заметки',
-            '2. Добавить новую заметку',
-            '3. Редактировать заметку',
-            '4. Найти заметку(и) по дате создания',
-            '5. Найти заметку(и) по дате изменения',
-            '6. Найти заметку по id',
-            '7. Удалить заметку',
-            '8. Закончить работу', sep = '\n')
-        return self.__input_choice()
-
-    def show_find_menu(self):
-        print('1. Определить заметку(и) по дате создания',
-            '2. Определить заметку(и) по дате изменения',
-            '3. Определить заметку по id',
-            '4. Вернуться к главному меню', sep = '\n') 
-        return self.__input_choice()
-
-    
-    def show_edit_menu(self):
-        print('1. Изменить название заметки',
-            '2. Изменить описание заметки',
-            '3. Отменить изменения и вернуться к главному меню',
-            '4. Сохранить изменения и вернуться к главному меню', sep = '\n') 
-        return self.__input_choice()
-
-
-    # def show_add_menu(self):
-    #     print('1. Ввести название заметки (обязательно для заполнения)',
-    #         '2. Ввести описание заметки (обязательно для заполнения)',
-    #         '3. Ввести дату создания (если дата заметки соответствует текущей дате, не редактируйте этот пункт)', sep = '\n') 
-    #     return self.input_choice()
-
-    def __input_choice(self):
-        try:
-            return int(input("Введите номер команды: "))
-        except:
-            print("Неверный ввод команды (ожидается целое положительное число)")
-            return 0
